optscale-deploy/ansible/roles/common/files/requirements.txt
docker==3.5.1
kubernetes==10.0.1
pyyaml==3.12
cryptography>=3.2.1,<=3.3.1
